\name{plotmo.methods}
\title{Please ignore}
\alias{get.plotmo.pairs}
\alias{get.plotmo.singles}
\alias{get.plotmo.y}
\description{Methods exported for use by \code{\link[earth]{earth}}.}
\usage{
get.plotmo.singles(object, env, x, trace, all1)
get.plotmo.pairs(object, env, x, trace, all2)
get.plotmo.y(object, env, y.column, expected.len, trace)
}
\arguments{
\item{all1,all2,env,expected.len}{See \code{plotmo.methods.R}.}
\item{object,trace,x,y,y.column}{See \code{plotmo.methods.R}.}
}
